Set up new file for exonPlot, initiated branching off
